requirements.txt
Flask==3.1.2
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
iniconfig==2.1.0
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
SQLAlchemy==2.0.44
Werkzeug==3.1.3
